Arreglado asociar tarjeta en Alta Cliente. Las tarjetas se muestran en baja cliente

+++ src/PagoElectronico/PagoElectronico/Model/Tarjeta.cs
@@ -10,13 +10,13 @@ namespace PagoElectronico.Model
         public int ID { get; set; }
         public int ID_Persona { get; set; }
         public int cod_seguridad;
-        public int numero;
+        public long numero;
         public string Emisor { get; set; }
         public DateTime Fecha_Emision { get; set; }
         public DateTime Fecha_Vencimiento { get; set; }
 
 
-        public Tarjeta(int id, int id_persona, int cod_seg, int num, string emisor,
+        public Tarjeta(int id, int id_persona, int cod_seg, long num, string emisor,
                        DateTime fecha_emision, DateTime fecha_vencimiento)
         {
             ID = id;
@@ -27,7 +27,7 @@ namespace PagoElectronico.Model
             Fecha_Emision = fecha_emision;
             Fecha_Vencimiento = fecha_vencimiento;
         }
-        public Tarjeta(int num, int cod_seg, string emisor)
+        public Tarjeta(long num, int cod_seg, string emisor)
         {
             numero = num;
             cod_seguridad = cod_seg;
